keep track of state to integrate angle; fix units; more precision on zeros

+++ imu.h
@@ -11,7 +11,7 @@ class IMU
     void calibrate();
     void read();
 
-    int16_t a_x_zero, a_z_zero, g_y_zero;
+    int32_t a_x_zero, a_z_zero, g_y_zero; // in 1000ths
     int16_t a_x, a_z;
     int16_t w; // deg/s * 10^-1
     LSM6 device;
